+++ scripts/build_network.sh
@@ -217,5 +217,4 @@ main() {
     package "$@"
 }
 
-
 main "$@"